% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractText.R
\name{extractText}
\alias{extractText}
\title{Extract Text from Image}
\usage{
extractText(imagePath, ocrLang = "eng")
}
\arguments{
\item{imagePath}{Path of image to extract text from.}

\item{ocrLang}{}
}
\value{

}
\description{
Uses \href{https://github.com/tesseract-ocr/tesseract}{Google's OCR} (optical character recognition) engine to extract text from images. This function is just a wrapper of multiple steps found in \href{https://docs.ropensci.org/tesseract/articles/intro.html}{Tesseract documentation}.
}
